requirements.txt
Flask==2.0.1
aiohttp
aiortc
av
google-auth==1.35.0
gunicorn==20.1.0
numpy==1.19.5
opencv-python-headless==4.5.3.56
pip==21.2.4
tensorflow-cpu==2.6.0
tensorflow-cpu-estimator==1.15.1
